Python 3.x compatibility
Some changes are trivial syntax or import differences, but other
are more subtle Unicode vs. bytes issues.

In many places, I don?t know what the encoding should be so I
just use 'ascii' in the hope that it will raise an exception
rather than behave wrong and produce mojibake.

Any suggestion on how to improve this is welcome...