Decode boundary before putting it in headers
`boundary` is a bytes variable, so the header ended up looking like this:
```
Content-type multipart/form-data; boundary=b'--------------GHSKFJDLGDS7543FJKLFHRE75642756743254'
```
The server then interprets the boundary as being "b'--------------GHSKFJDLGDS7543FJKLFHRE75642756743254'" and never finds it in the form data, so fails (silently) at reading the multipart.

We need to decode boundary into ascii before we can use it in the header so that it becomes
```
Content-type multipart/form-data; boundary=--------------GHSKFJDLGDS7543FJKLFHRE75642756743254
```

Debugged and tested locally with python 3.5.2. and sphinx-server==0.1.0